chore: partially repair storybook
