+++ botfront/imports/ui/components/stories/common/style.import.less
@@ -288,7 +288,7 @@
     border: 1.5px dotted;
     vertical-align: baseline;
     margin: 0em 0.75em 0em 0em;
-    padding: 0.78571429em 1.5em 0.78571429em;
+    padding: 0.628571432em 1.5em 0.78571429em;
     text-transform: none;
     line-height: 1em;
     font-style: normal;
@@ -296,6 +296,7 @@
     text-decoration: none;
     border-radius: 0.28571429rem;
     font-weight: normal;
+    font-family: @fontFamily;
 }
 
 .dashed:focus {
